refactor: useQuery 컴포넌트 분리
