+++ src/hooks/queries/queryKeys.ts
@@ -1,3 +1,4 @@
 export const QUERY_KEYS = {
   guestBook: (invitationId: string) => ['guestbook', invitationId] as const,
+  invitation: () => ['invitation'] as const,
 };
